Move components/common to shared/ui
