+++ src/shared/ui/slider/Slider.tsx
