+++ src/shared/ui/slider/slider.test.tsx
